Updated ClassFinder
ClassFinders:
Added an option to set the JarFile
Added an option to set the ClassLoader

This is useful when working with jar files loaded as modules. For example, Spigot plugins.

Fixed an issue where the IdeClassFinder wouldn't correctly find classes if the base package is empty.

Fixed UpToDate check, this will no longer say an update is available when the current version matches upstream.